Removed client component and made page client-rendered
